+++ CHANGELOG.md
@@ -5,6 +5,11 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+
+## **WORK IN PROGRESS**
+
+* (theimo1221) Update packages
+
 ## 3.0.0-alpha.84 (2024-09-07)
 
 * (theimo1221) Add possibility to add a heating rod to the dachs